==> code/02_html_parsing.ipynb <==
# Parsing HTML documents dowloaded from Proquest

1. Select articles you want to turn into a csv file (I assume that you want to downlaod only full texts). 
2. Print these selected articles. 
3. Save the print version of the articles as an HTML file. 
4. Repeat this process until you've downloaded all of the target files. 
5. Put these HTML files and this code in the same directory. 
6. Run the script. 

## Create a function

def parsing_proquest(x):

    # load libs 

    from bs4 import BeautifulSoup

    import re

    # load files
    
    soup = BeautifulSoup(open(x,"r"), 'html.parser')

    # save filtered results to new objects 

    doc_text = soup.findAll("text")
    doc_date = soup.findAll("", {"class": "abstract_Text col-xs-12 col-sm-10 col-md-10 col-lg-10"})

    # for loop over pages
    
    ## text 

    sum_text = []

    for i in range(len(doc_text)):
       sum_text.append(doc_text[i])

    ## year

    sum_date = []

    for i in range(len(doc_date)):
       sum_date.append(doc_date[i])
    
    
    # check 
    
    print(len(sum_text), len(sum_date))

    # combine the results as a list

    newspaper_list = {'text': sum_text, 'date' : sum_date}

    # return
    
    import pandas as pd 
    
    return(pd.DataFrame(newspaper_list))

## Load files 

# get working directory 

import os 

os.chdir('/home/jae/muslim_newspapers-selected/full_version/')


## For loop

# for loop over entire page results 

n = 0

temp_dataset = []
for filename in os.listdir(os.getcwd()):
    if filename.endswith(".html"):
        n  = n + 1
        print("file",n, filename)
        temp_dataset.append(parsing_proquest(filename))
            

## Turn into a Dataframe

# turn list of dataframes into a single dataframe 

import pandas as pd

newspaper_dataframe = pd.concat(temp_dataset)



## Save as a CSV file

# save the dataframe as a csv file

newspaper_dataframe.to_csv("/home/jae/muslim_newspapers-selected/full_version.csv")
