==> code/05_01_preprocessing_text.ipynb <==
# Text cleaning 

[Carlos Ortiz](https://www.linkedin.com/in/carlosortizdev/), [Sarah Santiago](https://www.linkedin.com/in/sarah-santiago-7a297b18a/), and [Vivek Datta](https://www.linkedin.com/in/vivek-datta/) did most of the coding. Jae Yeon Kim reviewed and modified the code slightly. Please use the `Python3` kernel to run this notebook.

## Import libraries

#Include relevant imports here

import re
import string
from nltk.corpus import stopwords
import nltk
nltk.download('stopwords')

# Data manipulation 
import numpy as np
import warnings
import pandas as pd
from pandas.api.types import CategoricalDtype
from sklearn.preprocessing import StandardScaler

# Data visualization 
%matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns

# ML
from sklearn.linear_model import LogisticRegressionCV, LogisticRegression
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction import DictVectorizer
from sklearn.decomposition import PCA

warnings.filterwarnings('ignore')

## Import data 

# Read in scraped articles from csv file to dataframe
sample = pd.read_csv('/home/jae/ITS-Text-Classification/raw_data/sample_articles.csv')
unlabeled = pd.read_csv('/home/jae/ITS-Text-Classification/raw_data/unlabeled_articles.csv')

articles = sample.append(unlabeled, ignore_index = True)

## Clean data


# Select relevant columns and convert date column into Datetime format

articles = articles[['text', 'source', 'group', 'date', 'intervention', 'expanding', 'distancing', 'assimilating']]

articles['date'] =  pd.to_datetime(articles['date'], format='%Y%m%d')

articles.head()

### Create a response variable 


# Adding column to determine if there is relevant info (expanding, distancing, assimilating) or not 

response = []

for x in articles['assimilating'].isnull().values:
    if x == True:
        response.append(0)
    else:
        response.append(1)

# Create new binary column called category based on this
# 1 indicates domestic issue, 0 indicates non-domestic issue

articles['category'] = response

articles.head()

### Check the balance of classes

To check whether we need to worry about imbalanced classification problems.

articles['category'].value_counts()

## Export cleaned corpus 

articles.to_csv("/home/jae/ITS-Text-Classification/processed_data/cleaned_text.csv")
