==> code/06_causal_inference.Rmd <==
---
title: "ITS design analysis"
author: "Jae Yeon Kim"
output:
html_document:
  toc: True
  theme: united
  number_sections: True
---

# Setup

```{r}
# Import libraries 
pacman::p_load(
        tidyverse, # the tidyverse framework 
        splitstackshape, # stacking and reshaping datasets 
        tsModel, # specifying time series regression models
        lmtest, # testing linear regression models
        Epi, # statistical analysis in epidemiology 
        splines, # spline regression
        vcd, # visualizing categorical variable 
        nlme, # linear and nonlinear mixed effects models 
        zoo, # S3 infrastructure regular and irregular time series 
        simpleboot, # bootstrapping 
        TTR, # constructing technical trading rules
        ggpmisc, # detecting peaks and valleys 
        broom, # tidying model objects 
        modelr, # modeling 
        forecast, # forecasting 
        ggseas, # seasonal and decomposition on the fly
        itsadug, # interpreting time series and autocorrelated data 
        mgcv, # mixed GAM Computation Vehicle with Automatic Smoothness Estimation
        tseries, # computational financial analysis
        mcp, # regression with multiple change points
        lawstat, # testing stats in biostat, public policy, and law
        sarima, # simulation and prediction with seasonal ARIMA models
        car, # Durbin Watson test 
        tidyr, # tidying messy daa 
        AICcmodavg, # computing predicted values and standard errors 
        MuMIn, # selecting models
        ggthemes, # fancy ggplot themes
        ggpubr, # arranging plots 
        stargazer, # regression table 
        here, # reproducibility 
        patchwork, # rearranging ggplot 
        purrr, # functional programming 
        testthat) # unit testing
```


```{r}
# devtools::install_github("jaeyk/makereproducible")
library(makereproducible)

# Import R scripts
script_list <- list.files(paste0(here::here(), "/functions"),
  pattern = "*.r|*.R",
  full.names = TRUE
)

for (i in 1:length(script_list))
{
  source(script_list[[i]])
}

# Publication ready plots 
ggplot2::theme_set(ggpubr::theme_pubr())
```

# Importing files

Unfortunately, I cannot share the original data because they are proprietary. ProQuest holds the copyrights. 

```{r include=FALSE}
# Import files 
sample <- readr::read_csv(here("processed_data/cleaned_text.csv"))[,-1]
unlabeled <- readr::read_csv(here("processed_data/predicted.csv"))[,-1]
```

# Renaming columns 


```{r}
# Rename column 
sample <- sample %>% rename(domestic = category)
unlabeled <- unlabeled %>% rename(domestic = category)
```


```{r}
# Select key variables 
sample_selected <- dplyr::select(sample, 
              source,
              intervention,
              date,
              domestic,
              text)

labeled_selected <- dplyr::select(unlabeled, 
              source, 
              intervention,
              date,
              domestic,
              text)

# Row bind the two dataframes  
df <- bind_rows(sample_selected, labeled_selected)

# Fix the coding error in the intervention variable (reversing values)
df$intervention[df$intervention == "post"] <- "1"
df$intervention[df$intervention == "pre"] <- "0"

df$intervention[df$intervention == "1"] <- "pre"
df$intervention[df$intervention == "0"] <- "post"
```

# Descriptive analysis

## Data wrangling

If you want to replicate the analysis, you can start from here using `df.csv` saved in `processed_data` directory..

```{r}
# Recode values in domestic variable 
df$domestic <- as.character(df$domestic) 
df$domestic[df$domestic == 1] <- "Domestic"
df$domestic[df$domestic == 0] <- "International"

# Create group variable based on newspaper names 
df$group <- ifelse(str_detect(df$source, "India"), "Indian Americans", "Arab Americans")

# Drop text column 
df$text <- NULL

# Save the processed data without the text column
write.csv(df, here("processed_data/df.csv"))
```

## Data visualization

### Raw data 

### No outliers and missing values 

```{r}
# Save plot 
  
cleaned_domestic_plot <- group_df(df, group) %>%
  filter(domestic == "Domestic") %>% 
  visualize_raw() +  
  facet_wrap(~group, scale = "free_y") + # facetting, y scale is free 
  ggtitle("Arab and Indian Immigrants' Domestic Political Interests") + 
  labs(tag = "A", caption = "")

cleaned_nondomestic_plot <- group_df(df, group) %>%
  filter(domestic != "Domestic") %>% 
  visualize_raw() + 
  ggtitle("Arab and Indian Immigrants' International Political Interests") + 
  labs(tag = "B") +
  facet_wrap(~group, scale = "free_y") # facetting, y scale is free 

cleaned_domestic_plot / cleaned_nondomestic_plot

ggsave(here("output", "cleaned_data_plot.png"), height = 8)
```


```{r}
cleaned_domestic_plot_source <- group_df(df, source) %>%
  filter(domestic == "Domestic") %>% 
  visualize_raw() +  
  facet_wrap(~source, scale = "free_y") + # facetting, y scale is free   
  ggtitle("Arab and Indian Immigrants' Domestic Political Interests") + 
  labs(tag = "A", caption = "")

cleaned_nondomestic_plot_source <- group_df(df, source) %>%
  filter(domestic != "Domestic") %>% 
  visualize_raw() +
  ggtitle("Arab and Indian Immigrants' International Political Interests") + 
  labs(tag = "B") +
  facet_wrap(~source, scale = "free_y") # facetting, y scale is free 

cleaned_domestic_plot_source / cleaned_nondomestic_plot_source

ggsave(here("output", "cleaned_data_plot_source.png"), 
       height = 8,
       width = 8)
```


# Interrupted time series design analysis 

## Base model
```{r}
# Create assignment variable 
df_grouped <- group_df(df, group)

# Divide data 

df_domestic <- df_grouped %>%
  filter(domestic == "Domestic") 

df_nondomestic <- df_grouped %>%
  filter(domestic != "Domestic")

save(df_domestic, df_nondomestic, file = here("processed_data", "base_its.RData"))
```


```{r}
# Visualize ITS analysis result for each subset of the data 
its_base_dom_plot <- visualize_base(df_domestic) + ggtitle("Arab and Indian Immigrants' Domestic Political Interests") + labs(tag = "A", caption = "")

its_base_nondom_plot <- visualize_base(df_nondomestic) + ggtitle("Arab and Indian Immigrants' International Political Interests") + labs(tag = "B")

its_base_dom_plot / its_base_nondom_plot

ggsave(here("output" ,"its_base_plot.png"), height = 8)
```

## Check for autocorrelation

The base model is naive. To make it more robust, first, I assess the correlation between the series and its time lags (autocorrelation). In both cases, ACF (correlation a time series and its lags) test shows there exits a weak seasonal trend. 

```{r}
# Create the full date 

all_days <- seq(min(df_grouped$date), max(df_grouped$date), by = "+1 day")

# Turn it into a dataframe 

all_days <- data.frame(date = all_days)

# Filling the missing days 

df_grouped <- all_days %>%
  merge(df_grouped, by = "date", all.x = TRUE)

# Divide data 

df_domestic_ts <- df_grouped %>%
  filter(domestic == "Domestic") 

df_nondomestic_ts <- df_grouped %>%
  filter(domestic != "Domestic") 

save(df_domestic_ts, 
     df_nondomestic_ts, 
     file = here("processed_data", "acf.RData"))
```

```{r echo=FALSE}
# Apply to each data
acf_dom <- acf_plot(df_domestic_ts) + ggtitle("Domestic Political Interests") + labs(tag = "A")

acf_nondom <- acf_plot(df_nondomestic_ts) + ggtitle("International Political Interests") + labs(tag = "B")

acf_dom / acf_nondom

ggsave(here("output", "acf_plot.png"))

```


## Parameterizing correlation

The new model parameterizes autocorrelation by accounting for the correlation between the time series data and its time lags. `corARMA() function` defines correlation structure. `p` argument specifies the autoregressive order, and `q` argument specifies the moving average order of the ARMA structure. I then find the combination of `p` and `q` that yields the minimum Akaike Information Criterion (AIC).

```{r eval=FALSE, include=FALSE}

pq_domestic <- mapply(correct_ac_domestic, c(1,2,3), c(1))

pq_nondomestic <- mapply(correct_ac_nondomestic, c(1,2,3), c(1))

pq_domestic 
pq_nondomestic 
```

Domestic 

P    1        2        3       
Q    1        1        1       
AIC  **3635.181** 3635.227 3636.323
BIC  3667.999 3672.733 3678.519
AICc 3635.322 3635.408 3636.55 

Non-domestic 

P    1       2        3       
Q    1       1        1       
AIC  2927.95 2923.072 **2899.469**
BIC  2959.07 2958.637 2939.48 
AICc 2928.13 2923.304 2899.759

## Visualization 

```{r}

# Visualize each result in the scatted plot 

its_adj_dom_plot <- visualize_adj(df_domestic, 1, 1) + ggtitle("Arab and Indian Immigrants' Domestic Political Interests") + labs(caption = "", tag = "A") +
  ylim(c(0,20))

its_adj_nondom_plot <- visualize_adj(df_nondomestic, 3, 1) + ggtitle("Arab and Indian Immigrants' International Political Interests") + labs(tag = "B") +
  ylim(c(0,20))

its_adj_dom_plot / its_adj_nondom_plot

ggsave(here("output", "its_adjusted_plot.png"), height = 8)
```

## Duration effect analysis

```{r eval = FALSE}
df_domestic <- df_domestic %>%
    mutate(year = extract_year(df_domestic))

df_nondomestic <- df_nondomestic %>%
    mutate(year = extract_year(df_nondomestic))

model_out <- map_dfr(c(1:6), model_spec)

save(model_out, file = here("processed_data", "duration_model.RData"))
```


```{r eval = FALSE}
model_out %>%
  mutate(Year = rep(2001:2006, each = 4)) %>%
  ggplot(aes(x = Year, y = Estimate)) +
    geom_point() +
    facet_grid(Category ~ Model)

ggsave(here("output", "duration_effect.png"))
```

## Regression table

### OLS

```{r eval = FALSE}

dom_output_ols <- lm(count_ts ~ intervention +  date + group, 
               data = df_domestic) 

nondom_output_ols <- lm(count_ts ~ intervention +  date + group, 
               data = df_nondomestic)

stargazer(dom_output_ols, nondom_output_ols,
          title = "ITS design analysis results",
          dep.var.labels = c("The publication count for Muslim-related articles on"),
          header = FALSE,
          single.row = TRUE,
          column.labels = c("Domestic politics",
                            "International politics"),
          covariate.labels = c("Intervention",
                               "Date",
                               "Indian Americans"),
          model.numbers = FALSE)

```

### OLS with robust SEs

```{r}
dom_output_robust <- lm_robust(count_ts ~ intervention +  date + group, 
                     data = df_domestic) 

nondom_output_robust <- lm_robust(count_ts ~ intervention +  date + group, 
                        data = df_nondomestic)

bind_rows(tidy(dom_output_ols, conf.int = TRUE) %>% mutate(model = "OLS"),
tidy(dom_output_robust) %>% mutate(model = "OLS with Robust SEs")) %>%
  filter(term != "(Intercept)") %>%
  ggplot(aes(x = term, y = estimate, 
             ymin = conf.low,
             ymax = conf.high)) +
    geom_pointrange() +
    scale_x_discrete(labels = 
      c("date" = "Date",
        "groupIndian Americans" = "Indian Americans",
        "intervention" = "Intervention")) +
    facet_wrap(~model) +
    coord_flip() +
    labs(x = "Term",
         y = "Estimate")

ggsave(here("output/ols_robust_ses.png"))
```
### GLS 

```{r}
dom_output_gls <- gls(count_ts ~ intervention +  date + group, 
               data = df_domestic,
               correlation = corARMA(p = 1, q = 1, form = ~ date | group)) 

nondom_output_gls <- gls(count_ts ~ intervention +  date + group, 
               data = df_nondomestic,
               correlation = corARMA(p = 3, q = 1, form = ~ date | group)) 

save(dom_output_gls, nondom_output_gls, 
     file = here("processed_data", "gls_base.RData"))
```

```{r}
stargazer(dom_output_gls, nondom_output_gls,
          title = "ITS design analysis results",
          dep.var.labels = c("The publication count for Muslim-related articles on"),
          header = FALSE,
          single.row = TRUE,
          column.labels = c("Domestic politics",
                            "International politics"),
          covariate.labels = c("Intervention",
                               "Date",
                               "Indian Americans"),
          model.numbers = FALSE)
```


### GLS with sourced fixed effects 

```{r}
pq_domestic_source <- mapply(correct_ac_domestic_source, c(1,2,3), c(1))

pq_nondomestic_source <- mapply(correct_ac_nondomestic_source, c(1,2,3), c(1))

pq_domestic_source # [1,1]

pq_nondomestic_source # [2,1]
```

```{r}
dom_output_fe <- gls(count_ts ~ intervention +  date + source, 
               data = group_df(df %>% filter(group == "Indian Americans"), source) %>% filter(domestic == "Domestic"),
               correlation = corARMA(p = 1, q = 1, form = ~ date | source)) 

nondom_output_fe <- gls(count_ts ~ intervention +  date + source, 
               data = group_df(df %>% filter(group == "Indian Americans"), source) %>% filter(domestic != "Domestic"),
               correlation = corARMA(p = 2, q = 1, form = ~ date | source))

save(dom_output_fe, nondom_output_fe, 
     file = here("processed_data", "gls_fe.RData"))
```


```{r}
stargazer(dom_output_gls, nondom_output_gls,
          title = "ITS design analysis results",
          dep.var.labels = c("The publication count for Muslim-related articles on"),
          header = FALSE,
          single.row = TRUE,
          column.labels = c("Domestic politics",
                            "International politics"),
          covariate.labels = c("Intervention",
                               "Date"),
          model.numbers = FALSE)
```

# Extract and export R code 

```{r eval=FALSE}
knitr::purl(input = here("code", "06_causal_inference.Rmd"), 
            output = here("code", "06_causal_inference.r"))
```
