==> code/07_additional_text_analysis.Rmd <==
---
title: "Additional text analysis"
author: "Jae Yeon Kim"
output:
html_document:
  toc: True
  theme: united
  number_sections: True
---

# Setup

```{r}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(
        tidyverse, # the tidyverse framework
        lubridate, # wrangling date and time
        ggpubr, # arranging ggplots
        ggthemes, # fancy ggplot themes
        tidytext, # tidytext
        patchwork, # arranging images,
        purrr, # functional programming
        here, # reproducibility
        httr, # httr connection 
        jsonlite, # parsing JSON
        glue, # gluing objects and strings         
        rio, # import and export files  
        forecast, # forecasting 
        stargazer
)

devtools::install_github("jaeyk/rnytapi", dependencies = TRUE)

library(rnytapi)

# Publication ready plots 
ggplot2::theme_set(ggpubr::theme_pubr())
```

```{r}
# Import R scripts

script_list <- list.files(here("functions"),
  pattern = "*.r|*.R",
  full.names = TRUE
)

walk(script_list, source)
```

# Collecting data using the NYT API

```{r eval = FALSE}
#Sys.setenv(nyt_key = "<insert key>")
key <- Sys.getenv("nyt_key")

# Parameters 

begin_year <- 1996:2005
end_year <- 1997:2006
term <- "muslim+muslims"
```

I need to split the date ranges to get around the NYT API policy on pagination (max 200 pages for a search result).

```{r eval = FALSE}
df1996 <- extract_all(term = 'muslim+muslims', begin_date = '19960911', end_date = '19970911', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df1997 <- extract_all(term = 'muslim+muslims', begin_date = '19970911', end_date = '19980911', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df1998 <- extract_all(term = 'muslim+muslims', begin_date = '19980911', end_date = '19990911', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df1999 <- extract_all(term = 'muslim+muslims', begin_date = '19990911', end_date = '20000911', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df2000 <- extract_all(term = 'muslim+muslims', begin_date = '20000911', end_date = '20010911', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')

df2001a <- extract_all(term = 'muslim+muslims', begin_date = '20010911', end_date = '20011231', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df2001b <- extract_all(term = 'muslim+muslims', begin_date = '20021231', end_date = '20030311', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df2001c <- extract_all(term = 'muslim+muslims', begin_date = '20020311', end_date = '20020911', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')

df2002a <- extract_all(term = 'muslim+muslims', begin_date = '20020911', end_date = '20021231', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df2002b <- extract_all(term = 'muslim+muslims', begin_date = '20021231', end_date = '20030311', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df2002c <- extract_all(term = 'muslim+muslims', begin_date = '20030311', end_date = '20030911', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')

df2003a <- extract_all(term = 'muslim+muslims', begin_date = '20030911', end_date = '20031231', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df2003b <- extract_all(term = 'muslim+muslims', begin_date = '20031231', end_date = '20040311', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df2003c <- extract_all(term = 'muslim+muslims', begin_date = '20040311', end_date = '20040911', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')

df2004a <- extract_all(term = 'muslim+muslims', begin_date = '20040911', end_date = '20041231', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df2004b <- extract_all(term = 'muslim+muslims', begin_date = '20051231', end_date = '20050311', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df2004c <- extract_all(term = 'muslim+muslims', begin_date = '20050311', end_date = '20050911', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')

df2005a <- extract_all(term = 'muslim+muslims', begin_date = '20050911', end_date = '20051231', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df2005b <- extract_all(term = 'muslim+muslims', begin_date = '20051231', end_date = '20060311', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df2005c <- extract_all(term = 'muslim+muslims', begin_date = '20060311', end_date = '20060711', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')
df2005d <- extract_all(term = 'muslim+muslims', begin_date = '20060711', end_date = '20060911', key = 'lJWFqmFXKLE8yoxN95xLjnyqMFk67a5I')

save(list = ls(pattern = "^df"), file = here("processed_data/nyt_parts.RData"))

df <- bind_rows(mget(ls(pattern = "^df")))

df <- df %>%
  # Filter non-successful search results
  filter(status == "OK") %>%
  # Remove duplicates 
  distinct()

write_rds(df, here("processed_data/nyt_articles.RData"))
```

# Munging data 

```{r}
df <- read_rds(here("processed_data/nyt_articles.RData"))

# Rename columns 
colnames(df) <- str_remove(colnames(df), "response.docs.")

nrow(df)
```


```{r}
# Select columns and filter rows 
df <- df %>%
  select(pub_date, section_name, subsection_name) %>%
  filter(!(section_name %in% c("Style", "Business Day", "Books", "Arts", "Magazine", "Movies", "Theater", "Sports", "Your Money", "Science", "Technology", "Food", "Health", "Real Estate", "Job Market", "Great Homes & Destinations", "Automobiles", "Fashion & Style", "Home & Garden", "Travel")))

nrow(df)
```


```{r}
# Format date column and rename it
df$pub_date <- lubridate::ymd(str_replace(df$pub_date, "T.*", ""))

df <- df %>% rename(date = pub_date)

write.csv(df, here("processed_data", "nyt.csv"))
```


```{r}
# Filter rows and mutate domestic column 
df <- df %>%
  mutate(domestic = ifelse(section_name == "World", "International", "Domestic"))

df$intervention <- ifelse(df$date > as.Date("2001-09-11"), 1, 0)

table(df$section_name)
```

# Data visualization

```{r}
domestic_plot <- ts_df(df) %>%
  filter(domestic == "Domestic") %>%
  visualize_raw() +
  ggtitle("Majority Interests in Domestic News on Muslims") + 
  labs(tag = "A", caption = "") +
  ylim(c(0,17))

nondomestic_plot <- ts_df(df) %>%
  filter(domestic != "Domestic") %>% 
  visualize_raw() +
  ggtitle("Majority Interests in International News on Muslims") + 
  labs(tag = "C", 
       caption = "Source: New York Times") +
  ylim(c(0,17))

domestic_plot + nondomestic_plot

ggsave(here("output/nyt_ethnic_raw_comp.png"), height = 8)
```
